app/pyproject.toml
[tool.poetry]
name = "nfc-app"
version = "0.1.0"
description = "NFCアプリ"
authors = ["TsubasaOura"]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
nfcpy = "^1.0.4"
customtkinter = "^5.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"